Tests/LibWeb/Text/input/external-object.html

<!DOCTYPE html>
<script src="include.js"></script>
<script>
    test(() => {
        println(`external.AddSearchProvider() returns undefined: ${external.AddSearchProvider() === undefined}`);
        println(`external.IsSearchProviderInstalled() returns undefined: ${external.IsSearchProviderInstalled() === undefined}`);
    });
</script>
