Tests/LibWeb/Text/expected/external-object.txt

external.AddSearchProvider() returns undefined: true
external.IsSearchProviderInstalled() returns undefined: true
